Migrate 'is-visible' test
